Move variantName util into process-new-page cloud function
